requirements/dev-mypy.txt
mypy==1.10.1
types-PyYAML
docutils-stubs
types-beautifulsoup4
types-jsonschema
types-appdirs
types-requests
types-clang
pytest
pydantic>=2.0
pydantic-extra-types
sphinx>=4.5
